Fix `typepigeon` version to pre 2 & test failures due to USGS response change (#60)
* Resolve issues with automated tests
- Fix typepigeon version to pre 2
- Fix polygon c'tor input shell array type not being 'number'
- Fix reference for updated hwm response
- Use better criteria for testing latest storm
- Update reference for USGS flood storms test as API response is updated
- Update flood event test refrence for updated USGS response
- Update usgs flood events test reference for updated USGS response
- Coops data within empty region explicitely return empty
- Documentation update & bugfix

+++ stormevents/nhc/track.py
@@ -821,7 +821,7 @@ class VortexTrack:
                                     row[["longitude", "latitude"]].values[None, :],
                                 ],
                                 axis=0,
-                            )
+                            ).astype(float)
 
                             quadrants.append(Polygon(vertices))
